projects/ETHcli/agent-evm/requirements.txt
# EVM Agent 依赖包
# 安装命令: pip install -r requirements.txt

# gRPC
grpcio>=1.60.0
grpcio-tools>=1.60.0

# LangChain 和 AI 模型
langchain-google-genai>=2.0.0
langgraph>=0.2.0
langchain-core>=0.3.0
langchain-community>=0.3.0
langchain>=0.3.0

# MCP 客户端
langchain-mcp-adapters



# 环境变量
python-dotenv>=1.0.0

# 其他依赖
pydantic>=2.5.0
